Finished the calculator project
Added the js for the buttons.
The operator buttons will only get event listeners after the first
number button is pressed.
The decimal button loses its event listener after the first click.
Regains the event listener when the "number" is done.
Consecutive operator button presses changes the operator to the last
pressed operator button.
A number followed by an "equals" operator simply outputs the number.
The "clear" button clears the arrays and variables and resets the flags
for decimals and operators.